move transpose out of sparse.py
