+++ python/tvm/topi/cuda/__init__.py
@@ -57,4 +57,5 @@ from . import tensorcore_alter_op
 from .argwhere import *
 from .scan import *
 from .sparse_reshape import *
+from .transfrom import *
 from .unique import *
